requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.2
drf-nested-routers==0.93.3
psycopg2==2.9.1
python-dotenv==0.19.0
pytz==2021.3
sqlparse==0.4.2
